Turn off fatal warnings in the Scala console
Warnings appear to have gotten stricter after we upgraded to 2.12;
instead of adjusting which warnings are enabled in the REPL after
each upgrade, it's more efficient to enable all warnings but stop
reporting them as errors. This results in more warning messages,
but at least the REPL will be usable, e.g. with unused imports.